+++ Frontend/.angular/cache/17.3.5/vite/deps/@angular_common_http.js
@@ -36,10 +36,9 @@ import {
   withNoXsrfProtection,
   withRequestsMadeViaParent,
   withXsrfConfiguration
-} from "./chunk-WJKSX2ZK.js";
-import "./chunk-CS6ADNSZ.js";
-import "./chunk-EKRORSVR.js";
-import "./chunk-QOHD3WUR.js";
+} from "./chunk-LHTDDO4X.js";
+import "./chunk-CVLI7GHA.js";
+import "./chunk-P6BN7I6E.js";
 export {
   FetchBackend,
   HTTP_INTERCEPTORS,
