+++ Frontend/.angular/cache/17.3.5/vite/deps/@angular_core.js
@@ -451,8 +451,7 @@ import {
   ɵɵvalidateIframeAttribute,
   ɵɵviewQuery,
   ɵɵviewQuerySignal
-} from "./chunk-EKRORSVR.js";
-import "./chunk-QOHD3WUR.js";
+} from "./chunk-P6BN7I6E.js";
 export {
   ANIMATION_MODULE_TYPE,
   APP_BOOTSTRAP_LISTENER,
