+++ Frontend/.angular/cache/17.3.5/vite/deps/@angular_platform-browser.js
@@ -35,13 +35,12 @@ import {
   provideProtractorTestingSupport,
   withHttpTransferCacheOptions,
   withNoHttpTransferCache
-} from "./chunk-2KUQ5AIC.js";
-import "./chunk-WJKSX2ZK.js";
+} from "./chunk-NQZXWTCR.js";
+import "./chunk-LHTDDO4X.js";
 import {
   getDOM
-} from "./chunk-CS6ADNSZ.js";
-import "./chunk-EKRORSVR.js";
-import "./chunk-QOHD3WUR.js";
+} from "./chunk-CVLI7GHA.js";
+import "./chunk-P6BN7I6E.js";
 export {
   BrowserModule,
   By,
