+++ Frontend/.angular/cache/17.3.5/vite/deps/chunk-NQZXWTCR.js
@@ -1,6 +1,6 @@
 import {
   withHttpTransferCache
-} from "./chunk-WJKSX2ZK.js";
+} from "./chunk-LHTDDO4X.js";
 import {
   CommonModule,
   DOCUMENT,
@@ -11,7 +11,7 @@ import {
   isPlatformServer,
   parseCookieValue,
   setRootDomAdapter
-} from "./chunk-CS6ADNSZ.js";
+} from "./chunk-CVLI7GHA.js";
 import {
   APP_ID,
   ApplicationModule,
@@ -42,6 +42,8 @@ import {
   Version,
   ViewEncapsulation$1,
   XSS_SECURITY_URL,
+  __spreadProps,
+  __spreadValues,
   _global,
   _sanitizeHtml,
   _sanitizeUrl,
@@ -67,11 +69,7 @@ import {
   ɵɵdefineInjector,
   ɵɵdefineNgModule,
   ɵɵinject
-} from "./chunk-EKRORSVR.js";
-import {
-  __spreadProps,
-  __spreadValues
-} from "./chunk-QOHD3WUR.js";
+} from "./chunk-P6BN7I6E.js";
 
 // node_modules/@angular/platform-browser/fesm2022/platform-browser.mjs
 var GenericBrowserDomAdapter = class extends DomAdapter {
@@ -1894,4 +1892,4 @@ export {
    * License: MIT
    *)
 */
-//# sourceMappingURL=chunk-2KUQ5AIC.js.map
+//# sourceMappingURL=chunk-NQZXWTCR.js.map
