+++ Frontend/package.json
@@ -19,6 +19,7 @@
     "@angular/platform-browser-dynamic": "^17.1.0",
     "@angular/router": "^17.1.0",
     "bootstrap": "^5.3.3",
+    "ngx-cookie": "^6.0.1",
     "ngx-cookie-service": "^17.1.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
